[move/#229] Transformation 파일 이동
