+++ feature/registration/src/main/java/com/napzak/market/registration/sale/component/ShippingFeeTextField.kt
@@ -26,8 +26,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.napzak.market.designsystem.theme.NapzakMarketTheme
 import com.napzak.market.feature.registration.R.string.sale_price_tag
-import com.napzak.market.util.android.adjustToMaxPrice
-import com.napzak.market.util.android.priceSeparatorTransformation
+import com.napzak.market.ui_util.adjustToMaxPrice
+import com.napzak.market.ui_util.priceSeparatorTransformation
 
 @Composable
 fun ShippingFeeTextField(
